requirements.txt
# Core audio and speech
pygame>=2.5.0
gtts>=2.3.0
sounddevice>=0.4.0

# Voice recognition
vosk>=0.3.0

# AI and API
openai>=1.0.0

# Arduino communication
pyserial>=3.5

# System utilities
pathlib2>=2.3.0

Adafruit-PCA9685
numpy
soundfile
colorama
